Move chapter 3 listings to standard file layout.
