--- src/main/kotlin/chapter3/Listing_3_8.kt
@@ -1,12 +1,0 @@
-package chapter3
-
-object Listing_3_8 {
-    /*
-    //tag::init[]
-    fun sum(xs: List): Int = when(xs) {
-        case Nil -> 0 <1>
-        case Cons(head, tail) -> head + sum(tail) <2>
-    }
-    //end::init[]
-     */
-}
