+++ src/main/kotlin/chapter3/tree.kt
@@ -1,13 +1,7 @@
 package chapter3
 
-//Tree structures
-//tag::init[]
 sealed class Tree<out A>
 
 data class Leaf<A>(val value: A) : Tree<A>()
 
-data class Branch<A>(
-    val left: Tree<A>,
-    val right: Tree<A>
-) : Tree<A>()
-//end::init[]
+data class Branch<A>(val left: Tree<A>, val right: Tree<A>) : Tree<A>()
