--- src/main/kotlin/chapter3/variadic_function.kt
@@ -1,10 +1,0 @@
-package chapter3
-
-object VariadicFunctionSidebar {
-    //tag::init[]
-    fun <A> of(vararg aa: A): List<A> {
-        val tail = aa.sliceArray(1 until aa.size)
-        return if (aa.isEmpty()) Nil else Cons(aa[0], List.of(*tail))
-    }
-    //end::init[]
-}
